chore: Service aware interface
This commit introduce a new interface meant to be implemented by all
services that provide a Service. The purpose of the interface is to
ensure that all service-related events are known to the implementers and
that those left un-implemented are done intentionaly.

When a plugin implements the interface, it will automatically register
the callbacks without having to explicitely declare them in
getHooksAndCallback method.

As of now, only MediawikiStandalone and Program are covered to limit the
scope, the work for the other services will be done in a dedicated commit.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: I3e31a06ad210acd4d6d063abdf19ebfbcfa3e98f