Run MediaWiki project instance upgrades when running `tuleap-cfg site:deploy`
Note this also try to run the upgrade script for project with service
enabled but that may not have been created already because the event has
not yet been processed. The MW update script does not throw an error in
this situation so this is not really an issue. This could be improved in
a future contribution once we know the exact status.

Change-Id: Ife3fb0b359faa5e575f1922502689b3f58be784d