Use article instead of li for followup comments
The followup comments of an artifact is displayed by using a `ul/li` set
of tags with css rules to remove bullet and default styling.

This causes regular use of `ul/li` in the comment body to have an
apparently wrong level of indentation: we should have discs instead of
circles.

Instead of unordered list, we now display a `section` of `article`s.

Closes request #27153: Do not use unordered list to display followup comments

Change-Id: If1a7b7ea3047220f6e679defa46076daf432ae9a