feat: parse commit message for each pushed commit
part of story #26802 have smart commit with Tuleap Git

How to test:
- forge your database to allow artifact closure in a git repository
- make multiple commits in the default branch (master or main)
- push the branch
- You should see debug lines in git_syslog like the following:
  [debug] Analyzing commit with hash <sha1> pushed by user #<user_id> in
project #<project_id>
  [debug] Found commit message <your commit message>

Notes:
For now, the "dispatcher" does the processing synchronously, but later
it will dispatch a message to a redis queue and the Processor will be
built by the corresponding event handler. This is why the dispatcher has
a "builder" object to build the processor dynamically, it will be used
later so it has been introduced now.

Change-Id: Ie7c2f6357f840af64365ba8d89c3a8c357b3a739