chore: request #27156: Remove POSTINSTALL.txt
Back in the day, when package was installed more or less by hand, Tuleap
would display the content of a file `POSTINSTALL.txt` at the root of the
plugin after its installation in the web interface.

Usage of this was limited to `docman` and `docmanwatermark`. Both files have been removed:
* git #tuleap/stable/630d3b7778f90159ca5cafae29b53313c140c9ef
* git #tuleap/stable/90e9b6a8ee99182cec5b1a38d05ac57dd60028cf

We can safely remove the feature.

Change-Id: I97f9adc5916c5d197662a0fde325cb59bad8e89e