feat: create a Git branch from REST API
This is part of story #26803 create Tuleap Git branch & pull request from artifact

A Git branch can be created in a Git repository if the user is able to write
in it with the REST endoint POST git/:id/branches.

The expected json body content is like:

{
  "branch_name": "string",
  "reference": "string"
}

The REST call does not return anything for now.

A branch cannot be created from a tag, only from another branch or a
commit.

The git command update-ref[1] has been used to do the branch creation.

To test, do not forget to deploy the new sudoers config file.

[1] https://git-scm.com/docs/git-update-ref

Change-Id: Ica42bfbfe801e53881658c9404204f1907641ceb