cleanup: Remove Properties view 🔥
User are encouraged to use the new Document UI, let's remove duplicate
feature in the legacy interface.

Properties details and edition of properties can be done directly in the
new interface.

Note: Docman_ItemAction* seems to be unused, they are removed.

Part of story #26808: have document part of Community Edition

Change-Id: Iaa68898c74b247ee78318bd99baccf7328458847