fix: no generation fuzzy or obsolete in .po
When we run `make generate-po`, generated `.po` files might contain
strings marked as fuzzy or obsolete. Those strings are not supposed to
be found in Tuleap codebase (which is not the case currently).

For a first step, I suggest to facilitate developer work by removing
automatically obsolete strings (they don't have to do manually the
purge) and by emptying fuzzy strings.

If we see that it is not enough we can later create a hook, or/and a CI
job do check that no fuzzy nor obsolete strings are contributed.

Apparently, usage of `msgattrib` changes the width of .po files, so
there is a bunch of files modified by this contribution.

request #27158: Do not generate fuzzy or obsolete strings

Change-Id: I5b122755df7a4aa2c7afd4489bf654c140c43957