feat: add context word to Reference match
part of story #26802 have smart commit with Tuleap Git

When you write any word before a reference (for example "closes"), that
word will be picked up by the regexp and will be output before the
replacement link.

Later, it will be used to see if a commit message should close a Tracker
artifact.

Change-Id: I752eee65bbfba7123699a98ea84536050d503b01