Only allow the "Tuleap" skin on the MediaWiki instances
If users choose another default skin in their preferences they can trap
themselves on MediaWiki without a way back to Tuleap.

Also, only the Tuleap skin will be tested so we should not allow the
others.

To test, redeploy the file (or rebuild the package and redeploy it).

Part of story #26042: install mediawiki standalone

Change-Id: I6fd355d9d4442f570ecb8ab5755ca546fe66c8c2