Disable all email features in MediaWiki
Using email features is challenging:
* MediaWiki must be configured to send the emails likle the rest of
  Tuleap
* Permissions must be verified so that a user that cannot access the wiki
  anymore does not receive the notifications

This is not something that can be done with the available data for the
moment so all the email releated features are disabled.

To test, either rebuild and reinstall the package or redeploy the file
manually. All email releated features should be disabled (e.g. the
special page `EmailUser/<user_id>`.

Part of story #26042: install mediawiki standalone

Change-Id: Ie04a02d20be97f02c1769f18b9ba4e5a4b8693ef