refactor: Uniform naming to enable/disable plugin
In the interface we have the capacity to install/uninstall and
enable/disable a plugin.

However in the code, for enable/disable we have many terms to describe
the same concept:
* enable
* avail
* available
* activate

Furthermore the interface have a dedicated view for uninstalled plugins
which is labelled "Available plugins". This does not help at all.

The proposal is to use enable/disable for plugin activation.

To ease the refactoring, I took the opportunity to add more typing in
PluginManager/PluginFactory.

No functional changes, you can still install/uninstall, enable/disable,
and `tuleap plugin:install` plugins.

Note: `tools/utils/admin/activate_plugin.php` was marked as deprecated
13 months ago, it's time to remove it.

Note: Web spaces and CGI are things of the past, no need to mention them
when disabling a plugin.

Note: In the db, the column `available` of the table `plugin` is kept as
is to not break future `git bisect`.

Closes request #27165: Uniform naming to enable/disable plugin

Change-Id: I666ddcd09875133a1f369338c6a67f8e8d71d44b