request #27169: Parse native return and properties types when building the OpenAPI specification file
Overall the quality of the information accessible in the API Explorer
should be improved and the REST endpoints should continue to work. Note
however that the validation might have been improved so incorrectly
formatted representations that were accepted previously might now be
rejected (as they should have been from the beginning).

Change-Id: I890488e0195604dad8c677e508a53ec0ef8cf363