feat(mediawiki): Suspend and resume projects
When a project is suspended or made active again, the corresponding
MediaWiki instance is suspended and resumed.

This approach assume that nothing went wrong in creation / suspend /
resume.

Worst cases: trying to suspend an instance that was not created will
fail but no data will be leaked (same idea for other combinations). A
step of robustness is likely to be needed on Instance events anyway.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: I22edd1be97a581392c2940b283bdf308c5cc9e51