feat(mediawiki): Activate service at project duplication
When a project is created from another, create corresponding MW instance
when relevant.

Nothing is done for XML activation because there is no template with
mediawiki standalone at this stage. From my quick look up, there is no
dedicated code ATM to handle "just and activation" because for most
other services, at XML import, there is a dedicated XML stuff for import
(to define who is admin for instance). It's likely that we will follow
this path when we will manage XML import.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: I1f123f0c4d9a92153552f3014ea42abb156c832d