Drop `PFUser::isLoggedIn` method
The role of the PFUser class is to hold user informations but it also
used to keep the "logged in" state per request. Mixing the two together
is a source issues when you start to have different way to be "logged
in" as Tuleap currently do because this is not determined by the user
informations but by informations transported by the currently processed
request. As a result, REST calls might incorrectly consider a user as
not logged in when in fact it has been properly authenticated.

Note this contribution does not solve the issue completly, methods like
`getSessionHash`, `getSessionId` and `isAnonymous` should also be
removed/moved out but at least we know have a proper and accurate way to
determine if the current user is logged in.

Part of story #26042: install mediawiki standalone

Change-Id: I8090a29450ea20e809b5c95cc753d35e141cfc2d