chore: Plugins that provides a service implement PluginWithService
The main intent of this contribution is to make it more consistent and
visible so next time someone has to create a new plugin that provide a
serivce, they won't loose days to look for all hooks that might be
needed.

As we all know that was copy/paste what was done elsewhere, this commit
is an attempt that this won't be forgotten.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: Ic3470b6726b51454ad2865c5e481ac460b47c090