fix: send a non updated route query when user want to sort a column
Part of story #26766 smooth usage of Document search

When clicking on a column to order it, the old route query can be sent
in addition to the column sort parameter.
For example, in some case, when a user just wants to order a column, the query sent may be `q=my_previous_search&sort=your_column`. While it should be only `sort=your_column`.

How to test the fix:
 - In the search box in the tree view, search for: azerty
=> You are redirected to the search view with the result (ok)
 - Erase the value of the Global Text field and apply
=> You see all the document result (ok)
 - Click on one column
=> The result should NOT be the combination of the global text "azerty" result and the sorted column. In the query param part of the address bar, you should have:
`sort=your_column`. And not `q=azerty&sort=your_column` anymore

Change-Id: I9c94e2cc31218c75e97d716b1323e1659ee000d7