Harden Tuleap MediaWiki PHP FPM SystemD unit
Add some additional limitation on what the process can write. In an ideal
world we could restrict exactly where the process is supposed to read
and write but this is not possible with the quite old version of SystemD
coming with CentOS/Red Hat 7.

To test: redeploy the unit, tell SystemD to do a `daemon-reload` and
restart the service. Everything should continue to work.

Part of story #26042: install mediawiki standalone

Change-Id: Ic3842eb59920ccb6b7ea11152d4f8ccc24f23ed4