refactor: replace postGitlabMergeRequest
part of request #27503 replace tlp-fetch + neverthrow by fetch-result

A small functional change is expected in the "create GitLab branch & MR"
artifact actions button. The error message is now concatenated.

No other functional change expected.

When creating a merge request, there should still be a success feedback
at the end. When it fails, there should be an error message at the top
of the modal. When the checkbox is not checked, there should be only one
feedback message.

Change-Id: Ib3f113d6a810fe051f8b0ffe8fa578f09517cec5