request #27507: Rewrite `sha1collisiondetector` tool in Rust 🦀
A crate porting the original C library to Rust makes it easy to move our
existing C code to a few lines of Rust.

This also give us the opportunity to move all the code inside the
monorepo instead of pulling it from an external repository.

To test, build the package and install it in your dev instance.
You should still be able to commit inside SVN repository.

We use an overlay that pull Rust from the binaries provided by the Rust
project itself. This allow us to use the target we want to get a static
binary without rebuilding the whole toolchain.

Change-Id: Ibe5332da4599598f21f0aab768665affa6052de0