refactor: remove usage of deprecated call to Artifact::createNewChangeset
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure:
- when the status semantic is not defined,
- when there are only open values (no closed values per semantic status
configuration),

In each case, when you push a commit in the main branch of a GitLab
repository integrated with Tuleap, and its configuration allows closing
artifacts, then a comment will still be added on the artifact to explain
the problem.

Notes:
There is still another call in the same class but it is not touched to
keep this contribution smaller.

Change-Id: I25a1e55588d37fb628e5195b913acae3116bba19