REST API is broken when the MediaWiki Standalone plugin is enabled in production
A REST representation of the plugin makes a references to a class
located in a tests directory. This is not shipped in production (and not
added to the autoloader). Due to that Restler is not able to properly
parse all the routes and it crashes without logging anything.

To prevent similar future issues this contribution makes sure production
code cannot reference "test" code thanks to Deptrac. Psalm baseline has
been re-created to accomodate with the changes related to the usage of
\DataAccessResult.

Part of story #26042: install mediawiki standalone

Change-Id: If49d88a9ff1390f697e20c7e1f143ad4d84af728