feat: Export/import allow_artifact closure in XML
Export a project with 2 repositories:
* one old (with allow_artifact_closure=0 in db)
* one new (with allow_artifact_closure=1 in db)

In the generated xml, you should have an attribute
allow_artifact_closure for each repository, and the value should match
the original settings.

Import in a new project. Open `plugin_git` table.
➡️ The allow_artifact_closure column should reflect what have been
imported.

Note: the change of Git_GitoliteDriver is needed to be able to run tests
in the IDE, which was complaining that required parameters should be set
before optional ones (linked to PHP 8.1?)

Part of story #26802: have smart commit with Tuleap Git

Change-Id: Iae39f123c380eb2401b77cb2755e816c82b7ffa1