Extract common part of OIDC token generation
OIDC Core and extensions like the Back-Channel logout share similarities
in the way the tokens they use are generated [0][1]. This contribution
extracts the common part so we can re-use it later to generate a logout
token.

This is a technical change, no change in behavior is expected.

Part of request #27519 TuleapFarm: unlog one user everywhere

[0] https://openid.net/specs/openid-connect-core-1_0.html#IDToken
[1] https://openid.net/specs/openid-connect-backchannel-1_0.html#LogoutToken

Change-Id: I966118d55689cd14b7064cf58e7e218c8387f5fb