fix: title attribute in search results
When document's title is long and the viewport is small, the title is truncated.

In the tree view, the full document title is available on hover of the truncated title.

It should be the same in the search results.

Fix request #27527: Have full document title on hover in search results

Change-Id: I4eff52d95a973254bf312afd2f8803acb8007c84