request #27535: Remove the feature flag of the cross report export
The feature flag was added in order to collect feedbacks and give us
some time to rework the configuration modal. The rework of the
configuration screen will not happen soon so we consider it is better to
give access to the features to users even if it is hard to use and has
UI quirks. A message has been added in the modal to warn the users about
this.

This reverts commit 224215d01b03dacc3bd135a528bf7b40ea82db35.

Change-Id: I327ea7a9959956e1c100f1bd1c4092a6ef4bffe1