refactor: remove second deprecated call to Artifact::createNewChangeset
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure:
- when there is a Done or a Closed value, the Artifact is closed with a
comment.
- when there is no Done or Closed value, or there is no Status field, a
comment is added on the artifact to explain the problem.

Notes:
This also replaces the string parameter by an interface  called
ArtifactClosingCommentInCommonMarkFormat. It is stricter than allowing
any string to be passed.

Change-Id: I949ac1d4ab8feab24ca78134642a19b85a2e60d1