Make `openid`, `email` and `profile` scope accessible even when the OAuth2 server plugin is not present
This is needed for the MediaWiki standalone plugin where those scopes
are required to log the user in and collect some information so we can
create its profile on the MW side.

Note the `offline_access` scope is still only provided by the OAuth2
server plugin. It is not needed for the MediaWiki standalone accesses
and it has far greater power (borderline dangerous) so we will see to
make it accessible to everyone if/when we need it.

Part of story #26042: install mediawiki standalone

Change-Id: I0c34be3d2a1625e1df51c6e8cee5459a575a0a17