Closes request #27548: Introduce Vitest as the new default JS test runner
This contribution comes with an ADR explaining the rationale behing this
choice and it introduces the necessary architecture to use it. One small
library is migrated as a demonstration. Other migrations will be handled
in dedicated requests.

Vite is upgraded to 2.9.14 from 2.9.1 to avoid issues with the types.
Changelog can be seen here:
https://github.com/vitejs/vite/blob/v2.9.14/packages/vite/CHANGELOG.md

Change-Id: I9c322b10660ac19183b131e2fe98e85e1770798b