fix: Order custom properties by their label
As stated in the original AC of story #24197, the extra columns in
search results should be ordered alphabetically.

As of today, the order is not guaranteed because the underlying query is
done without sort. It means that the order is driven by the database and
can even change from one page load to another.

The order of column is now:
* Id (always there)
* Title (always there)
* Description
* Owner
* Create date
* Update date
* Status
* Obsolescence date
* Location
* Filename
* Custom properties **(order by label)**

For consistency, custom properties order is applied also in properties
administration and in update item properties modal.

Closes request #27530: Search results columns should be ordered alphabetically

Change-Id: Ia7ccb63f6dcef07f444fb0deccc853a04d602e7a