refactor: Remove dependency to PostPushCommitWebhookData
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure:
- when there is a Done or a Closed value, the Artifact is closed with a
comment.
- when there is no Done or Closed value, or there is no Status field, a
comment is added on the artifact to explain the problem.

Notes:
The ArtifactUpdater now returns a specialized Fault to indicate the
artifact is already closed. The caller is now responsible for logging
the info message.
ArtifactUpdater is now responsible for catching
NoPossibleValueException. It turns it to a Fault that is then logged as
an error, like other faults.

Change-Id: I319b5a6a37d90b6125740a610d39864a5bd14f07