refactor: use an interface for the Bad semantic comment
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure:
when there is no Done or Closed value, or there is no Status field, a
comment is added on the artifact to explain the problem.

Notes:
The comment metadata are not very important for the caller, so they are
hidden in PostPushCommitArtifactUpdater to make it easier to call. The
caller provides two comment bodies that are always in CommonMark format,
and ArtifactUpdater provides the metadata.

Change-Id: I8e80d481d8e1980536e573149d635c74b01f022a