refactor: move ArtifactUpdater to Tracker plugin
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure:
- when there is a Done or a Closed value, the Artifact is closed with a
comment.
- when there is no Done or Closed value, or there is no Status field, a
comment is added on the artifact to explain the problem.

Notes:
Moving it to Tracker is needed so that Git plugin will be able to
dispatch an event to call it. Git plugin must not depend on GitLab
plugin being present, but it can expect Tracker plugin to respond to the
event. After all, if there is no Tracker plugin, there is nothing to
close.

Change-Id: If6af2c9327bd6bcd44a0ed8909a347f5af3f00e6