feat: Enhance feedbacks at Git branch creation
This is part of story #26803 create Tuleap Git branch & pull request from artifact

When the branch is created, a feedback is displayed with direct links to
the created branch and to the Git repository.
When there is an error, a translated message from the REST API is now
displayed in the modal to add information for the user about the error.

Change-Id: I2c757d949e44ecee92b17da7884a6ce0986b9579