Closes request #27552: Add export of report columns as an XLSX spreadsheet
The feature re-uses the existing code of the cross-report export, the
document generated is exactly the same as the one you get when you open
the cross report export modal and directly click on export. All the code
added by this contribution is only technical to start the download when
the user clicks on the link.

Change-Id: Ib4bc63251d5f45ee9078d7cfeba7fdfb28beaa79