chore: HTML URL of a branch must be provided in the REST representation
This is part of story #26803 create Tuleap Git branch & pull request from artifact

The branch HTML URL format is relative to match other Git HTML URLs
provided in REST representations.
The REST route POST git/:id/branches now returns a GitBranch
representation.
This representation is now used in the modal to get the HTML URL of the
created branch in the feedback.

Change-Id: Id4ae2c35f167389b7c534d212b8f2bc137cf5a35