refactor: extract dedicated class for closing keyword
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure: If you push a
commit with a "close" keyword and a reference to a Tuleap artifact (for
example TULEAP-6555), the artifact is closed with a comment.

Notes:
We will need to have a way to parse the closing keywords for Git
commits. This change introduces a dedicated class for this purpose. In a
next contribution, it will be moved to Git plugin.

Change-Id: I6d7f902bdc15b8fcbd2a3880e9711b34076e16e4