refactor: move ClosingKeyword to Tracker plugin
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure: If you push a
commit with a "close" keyword and a reference to a Tuleap artifact (for
example TULEAP-6555), the artifact is closed with a comment.

Notes:
Instead of moving it to Git plugin, it is moved to Tracker plugin
because it has strong coupling with Artifact notions. Git plugin does
not know what an Artifact is, and it should not know it. It does not
depend on Tracker plugin after all.

Change-Id: I94581bb60b40cb5930b8a16ec0ec7734e47e5bc8