feat: Create a pull request from artifact
In the actions button of an artifact, the "Create Git branch" entry now
allow users to associate a pull request to the new branch.
The option is checked by default.

Note: the label of the dropdown entry will be changed later in a
dedicated commit.
Note: the PR creation should not be available if `pullrequest` plugin is
not enabled. This will be done in a dedicated commit.

Part of story #26803: create Tuleap Git branch & pull request from artifact

Change-Id: I0798ad05cf443f51a72337a7a86cb1448ed1b41b