feat: Adjust "Create Git branch" label
The artifact actions dropdown allows to create a git branch and a pull
request. The label should reflect that, like it is done for gitlab
entry.

However the label (and subsequent PR creation) depends on the
availability of the plugin pullrequest:
* If the plugin is not installed, or disabled, then the label is "Create
Git branch" and there isn't any checkbox in the modal to allow creation
of the PR.
* If the plugin is enabled, then the label is "Create Git branch and
pull request" and the checkbox is present.

From the implementation point of view, I took the opportunity to convert
the constant based event to an object, which should not have any
repercussion on existing code, like retrieving git/:id/pullrequests.

Part of story #26803: create Tuleap Git branch & pull request from artifact

Change-Id: I0938237e023b2bd145570c0fb2103905ac09e428