refactor: Drop @tuleap/tlp-fetch in favor of @tuleap/fetch-result
part of story #26803: create Tuleap Git branch & pull request from artifact

No functional change expected in the "Create Git branch" modal, the list
of repositories is still retrieved.

Notes:
One fewer dependency.
We should probably display the error message of that request somwhere,
it was not clear how to deal with it with the existing code so I chose
to default to an empty array (which prevents from selecting anything).
We should do that in a dedicated contribution.

Change-Id: Ie9ca3fdcf868eb0698ebf0d371ce62f69b7a1d9b