feat: find reference with closing keyword
part of story #26802 have smart commit with Tuleap Git

How to test:
- forge your database to allow artifact closure in a git repository
- commit on the default branch (master of main) of the git repository.
Your commit message should have a "closing reference" to an artifact in
the same project, like the following:
  closes art #123
- push the default branch
- You should see a debug line in codendi_syslog like the following:
  [debug] Found reference art #123 with closing keyword

If you write a reference to anything other than an artifact (for example
doc #123, custom #123), it is silently ignored.
If you write a reference to an artifact from another project, it is
silently ignored.

Change-Id: I588378400b30883db720e2478249fac27c2bc6c0