request #27558 New Project service Enable Button is greyed out
When creating a new custom service, the Enable button is greyed out, and when you finish creating the service, it is not enabled.

The expected result is to have a functionnal Enable button when creating a new custom service from the Project Administration Page. Currently, this enable button is greyed out which means it can't change values. Furthermore, the newly created custom service isn't enabled.
Currently :
  - Create a new custom service
  - Enabled button is greyed
  - Service is not enabled after creation.

To test :
  - Create a new custom service
  - Enable button should not be greyed out and service is directly enabled after creation

Change-Id: I2faad70cab59bf8c03cbfcc6a663a29da73de77e