refactor: interface before origin reference
part of story #26802 have smart commit with Tuleap Git

No functional change expected in GitLab artifact closure: when there is
a Done or a Closed value, the Artifact is closed with a comment with a
reference pointing back to the GitLab commit.

Notes:
By extracting an interface for this reference, we allow Git plugin to
give a reference to a Git commit in the event to trigger the closure of
the artifact. We cannot give the entire comment because Git does not
know about Tracker Artifacts (parsing of references is done in Tracker
plugin). This way, Git plugin does not know about Tracker Artifacts and
Tracker plugin does not know about Git commits, but they can work
together.

Change-Id: Ic609adb0a3292bd91bbd75083c07223ecec551e2