feat: Close artifacts referenced in commit message
part of story #26802 have smart commit with Tuleap Git

How to test:
- forge your database to allow artifact closure in a git repository
- commit on the default branch (for example master or main) of the git
repository. Your commit message should have a "closing reference" to an
artifact in the same project, like the following:
  closes art #123
- push the default branch
- You should see a debug line in codendi_syslog like the following:
  [debug] Closed artifact #123
- The artifacts you referenced are closed with a follow-up comment (from
workflow manager) referencing the @user_login, the closing keyword (same
as GitLab) and a reference to the Git commit. The @login should show
your user on hover and the reference should redirect to the commit you
pushed.

If you write a reference to an artifact that does not exist or that you
don't have permission to see, it is silently ignored.
If you write a reference to an artifact from another project, it is
silently ignored.
If the closing comment does not match the allowed keywords, it is
silently ignored.
If there is no Done semantic, it will fall back on Status closed value.
If there is no Status closed value (all values are open or semantic is
not set), it will add a follow-up comment with a message (similar to
Gitlab).

Notes: I am adding lines to psalm baseline because I don't know how
risky it is to add types to UserManager::getUserById(). It has extremely
high usage in the codebase.

Change-Id: I6ad4230979bc83e69ff2988bc4c4895bfad0c37b