Revert "chore: Switch emoji-button 4.6.2 to picmo 5.2.0"
This reverts commit 8f8bdf60ee50b4cdfb5bd88c94297f6dd850a9cf.

Library has decided it was fine to pull data at runtime from a third
party server by default. This is not fine for our use cases (and
possibly for anyone affected by GDPR).

This revert the update of the library, to be merged back the following
elements needs to be taken into account:
* no external queries
* the InMemoryStore should be used
* the messages and emoji data should be loaded from the appropriate
  locale

Part of request #27554: Switch emoji-button 4.6.2 to picmo 5.2.0

Change-Id: I36b758e9d18959f8e05dd92468269b14c1f91fbc