feat(mw-standalone): Add Mediawiki Standalone into Docker images
As Tuleap has to connect to itself for managing MediaWiki instances
(HTTP call) the certificate exposed by nginx is automatically added to
the trust store.

This commit is also an opportunity for a small refactoring of
Supervisord class in order to reduce the duplication between TCE and TEE
images.

At this time, the commit can only be tested with TCE because no TEE
images embed MW standalone stuff. TEE is expected to -partially- fail
because supervisord will attempt to start php74 fpm will the package has
not been installed yet.

Close request #27891 Mediawiki Standalone should be available in docker images

Change-Id: Id73f16f71423cedd8b09a778a1ad32a0369ed5c4