request #27893 Ldap: nightly syncro does not work properly
During ldap nightly syncro with ugroups, if a user have already been searched for a previous group, he will be removed from the other groups.

How to test :

    have a ldap group with 2 users
    have two different groups. Bind them to the ldap group with daily synch but without preserve_members option : the two users are in both groups
    run tuleap daily-job ==> users in one group have disappears.

During the daily synchro, the LDAP_GroupManager is the same for all
groups. So, if a user have been searched for another group, he is still
in the cache. We need to clear this cache after every group sycnhro.

Change-Id: I30b4302e51cb9dd6676372c509757297a51d9ffa