fix: request #27897 Welcome mail to new user can have multiple languages
Given a site administrator that creates a new user from the global administration interface,
if its Tuleap language is not the same as the platform default language (or en_US if not defined),
then the welcome mail received by the user will have multiple content languages.

The title is translated using the site administrator defined language,
and the body content of the mail uses the platform default language (or en_US if not defined).
Both title and body contents must use the site administrator used language.

Change-Id: I83bed9102e2c23a0ebac5d12472f00b599fc3e34