Package `tuleap-mercure` binary
This contribution packages the `mercure` binary so it can be installed
on Tuleap instances.

Using the pre-built binary provided by the upstream project is not
adequate because it does not give us visibility on what dependencies we
are injecting into our codebase and the current version is bundled with
an outdated version of Caddy.

To test, you can do something like this:
1. Build the RPM with `nix-build ./src/additional-packages/tuleap-mercure.nix`
2. Copy the resulting RPM into your dev container and install it
3. Check that you can execute `/usr/bin/tuleap-mercure`, you can start
   the whole server with a Caddyfile looking like this:

```
{
	auto_https off
	admin off
}

http://localhost:3000

log

route {
	mercure {
		transport_url bolt:///tmp/mercure.db
		publisher_jwt ChangeMe HS256
		subscriber_jwt ChangeMe HS256
	}

	respond "Not Found" 404
}
```

Deploying a systemD unit is left out for future contributions.

Part of story #27884: have access to a minimal working mercure environment

Change-Id: I0b55270c11860f5988bd2d62a1610b42847fcab2